44 console fix

+++ Task04/44_NUMBER ARRAY SUM/Program.cs
@@ -8,8 +8,10 @@ namespace _44_NUMBER_ARRAY_SUM
         static void Main(string[] args)
         {
             //var myArray = new char[] { '1', '3', '0', '7', '9', '5' };
-
-            var myArray = new int[] { 1, 3, 0, 7, 9, 5 };
+            //var myArray = new float[] { 1, 3, 0, 7, 9, 5 };
+            //var myArray = new double[] { 1, 3, 0, 7, 9, 5 };
+            var myArray = new long[] { 1, 3, 0, 7, 9, 5 };
+            //var myArray = new int[] { 1, 3, 0, 7, 9, 5 };
 
             Console.WriteLine(myArray.Sum());
         }
